feat: unipipe terraform can handle instances without binding
unipipe terraform requires a binding to actually execute terraform code. But the flow in a marketplace is that first an instance is created and once it is successfully provisioned, the binding can be created. With this new behavior, an instance without bindings will simply be set to succeeded, so bindings can be created for it.