fix: local commit detection not thread safe
synchronizeWithRemoteRepository did not handle the scenario where other
threads may modify the local repository between the merge operation and
when hasLocalCommits is set to false. This can cause the local view of
the repo to become out of sync with the remote. At best it is resolved
when another OSB operation, e.g. a provision request, creates a new
commit with the side effect that it signals there are local changes. At
worst the commits are lost if the repo storage is ephemeral and
is destroyed before the commits are pushed.