feat: remove terraform-runner docker image
with unipipe 2.0 we plan on dropping support for the terraform runner
image published together with unipipe. Users that still want to leverage
the docker image are welcome to fork the Dockerfile and maintain their
own build. The "unipipe terraform" command in unipipe cli continues
to be supported at this point.

As an OSS project it's important that we focus our resources on
maintaining widely used functionality and keeping up with the
terraform/opentofu ecosystem requires constant maintenance which we are
not able to offer at this point.